requirements.txt
fastapi==0.111.0
groq==0.8.0
uvicorn==0.30.1
python-dotenv==1.0.1

torch==2.0.0
torchaudio==2.0.1
faster-whisper==1.0.2
transformers
pandas==1.4.2
setuptools>=65.0.0
nltk==3.7



